Add native check support and tests for preview version (21)
